clio/relation_service/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
drf-spectacular>=0.27
django-filter>=24.0
psycopg[binary]>=3.1
gunicorn>=22.0
uvicorn>=0.30
